to issue#11762: fix review
